fix: keyboard collision, WASM over-read, image-load overflow, spinner hang
Review-driven quick wins across web and SDK:

- web: plain R was handled by two window keydown listeners at once —
  gizmo scale mode (CompositeViewport) and reset-rotation
  (useKeyboardShortcuts). Reset-rotation now lives on Shift+R, the gizmo
  handler ignores shifted keys, and the shortcuts overlay documents the
  W/E/R gizmo keys plus the new binding.
- web: estimateLighting resolves with neutral defaults on image decode
  failure instead of leaving the promise pending (upload spinner hang).
- web: remove dead, never-imported SurfacePlanes component.
- sdk: depth_wasm_render / DepthSceneJS::render copied width*height*4
  bytes from a composite sized to the background, over-reading the heap
  when dimensions mismatch; both now fail explicitly on mismatch.
- sdk: Image::load / load_from_memory computed the pixel byte count in
  int, overflowing for images past ~23 MP; now size_t.
- lockfile metadata sync from npm install.

Verified: 81/81 unit tests, lint, production build, 7/7 Playwright e2e,
plus live browser check of the new key bindings.

+++ sdk/src/io/image.cpp
@@ -78,7 +78,7 @@ Image Image::load(const std::string& path, Status* status) {
     img.width_ = static_cast<uint32_t>(w);
     img.height_ = static_cast<uint32_t>(h);
     img.format_ = PixelFormat::RGBA8;
-    img.pixels_.assign(data, data + (w * h * 4));
+    img.pixels_.assign(data, data + (static_cast<size_t>(w) * h * 4));
     stbi_image_free(data);
 
     if (status) *status = Status::Ok;
@@ -98,7 +98,7 @@ Image Image::load_from_memory(const uint8_t* data, size_t size, Status* status) 
     img.width_ = static_cast<uint32_t>(w);
     img.height_ = static_cast<uint32_t>(h);
     img.format_ = PixelFormat::RGBA8;
-    img.pixels_.assign(pixels, pixels + (w * h * 4));
+    img.pixels_.assign(pixels, pixels + (static_cast<size_t>(w) * h * 4));
     stbi_image_free(pixels);
 
     if (status) *status = Status::Ok;
